requirements.txt
Unidecode==0.04.16
corenlp-python==3.2.0-3
pexpect==3.2
wsgiref==0.1.2
xmltodict==0.9.0
